requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
azure-ai-projects
azure-identity
sqlalchemy
pymysql
python-multipart
alembic
